refactor: compnents에서 _shared로 폴더 이동
